requirements.txt
jwt~=1.3.1
requests~=2.28.2

websocket~=0.2.1
websockets~=11.0
websocket-client~=1.5.1
rel~=0.4.8
openai==0.27.4
rsa~=4.9

sanic==23.3.0
aiohttp~=3.8.4
pycryptodomex==3.17